glados: t/a/shoot.ts Don't flicker out of CatchBall
In t/a/shoot we use catchBallNecessary to decide whether to use CatchBall,
which continuously updated the catchPosition or to keep our current estimate
of the catch Position for more stability.
However, the decision whether to use CatchBall or not mostly depends on the
time difference the Robot and the Ball have to arrive at their destinations.

Since CatchBall in most cases returns a CatchPosition where the robot arrives
before or at roughly the same time as the ball, CatchBall would only be active
for a single frame in most cases.

Therefore we do not make proper use of the more precise (but less stable)
CatchBall.