glados: t/a/stopattack: simplify pass covering angle code
To be honest, I am not sure what the code actually did before. However,
what it intended to do (see #324 and #475) was to sometimes steal
opponent passes after their freekick, but to not stray to far from the
StopAttacker's original position (i.e. the one defending the ball ->
goal line)

We now allow a slightly wider angle to steal passes. This should be fine
as long as we are far enough away from the goal, and our goalie can
actually drive.