glados: a/a/feintgoalshot: Be less sensitive to tracking lookahead
The ball may be projected forward quite a bit when using a large system
or command delay (e.g. 30ms/70ms at RoboCup 2024). This can lead to the
strategy assuming the ball has already stopped and the volley has
failed.

Waiting 60 more milliseconds before deciding this should hopefully be
short enough to still be responsive, and long enough to not be too
sensitive.