glados: te/m/feintkeepertest.ts Introduce automated keeper testing
This test move lets you manually place the ball on the field.
Once a force start command is issued the attacking robot will
approach the ball and first look into one corner of the goal
to lure the keeper there.
Then it will shoot into the other corner.

This move can be used to test control parameter, keeper code
or anything of the kind.

After the ball either enters the goal or exits the defense
area, a stop command is automatically given (requires debug
mode to be on) to allow for a human to place the ball again.