amun/radiosystem: change semantic of a constructed transceiver to be that of an open transceiver
This is a major refactor of the Transceiver2015 class and introduces the
TransceiverHBC class to be able to handle differences in their interfaces
more easily in the future.
The idea is to make the ownership relationship between RadioSystem, USBThread
and TransceiverLayer simpler to reason about.
To be able to do this the constructors of Transceiver2015 and TransceiverHBC
were made private and only constructable via a static method acting as a
factory, which handles making sure that the Transceiver is actually connected
before returning and being able to connect to multiple transceivers of the same
type.
Some functions were changed to return an optional<TransceiverError> instead of
directly emitting the errors to be able to decide whether e.g. handling the
initial connection happened completely without errors.