ra: fieldwidget: visualize corner blocks
Commit `74afb1f4f37a6d40cc9760f2c4807ad1cdc66d89` implemented
the corner blocks in the simulator. This commit visualizes them
in the field widget.

This required an explicit setter in the fieldwidget, as the
simulator converts the geometry to an SSL geometry packet,
which doesn't contain information about the corner blocks.

This way the information about them is lost and needs to be
transferred another way.