amun: processor/tracking: Parse SSL_WrapperPacket protobuf just once
Parsing it multiple times is a bit wasteful. Also, parsing in the
processor allows passing the geometry to a new module, which will be
added later, more easily.

Technically, the semantic is slightly different now, in that
m_hasVisionData will not be set to true for invalid vision packets now
(since the parsing fails early, instead of in Tracker::process).
However, this was questionable behavior and probably does not matter.