amun: processor: Extract field parameters, geometry and ball model to a distinct class
This improves separation of concerns by having the Tracker concern
itself only with world state (i.e. where robots and the ball are)
without polluting the Processor.

This fixes an issue introduced by 3ee59a67b0b9. It aimed to factor out
the division calculation from assembleStatus, but also took along the
ball model adding code. Because of this, the model was not included in
the strategy status. Since the geometry including the ball model is now
added through WorldParameters in assembleStatus, the strategy status now
contains the ball model again.

Non-Processor users of WorldParameters set isReplay in the constructor
to true to prevent the ball model being written in such cases. This
should not matter, as the ball model is only written on updates and
probably no such non-processor user updates the ball model, but it does
not hurt.

The boolean fields in Processor are moved to the top to prevent an
initialization order problem, since they are needed during the
WorldParameters construction.