cmake: Add intermediate INTERFACE library for ExternalProjects
While we add an imported library for each ExternalProject and the Cmake
Documentation for add_dependency [1] specifies

> Dependencies added to an imported target or an interface library are
> followed transitively in its place since the target itself does not
> build.

our ExternalProjects where sometimes not built before their dependents
(especially when using make with just a single job processor).

While adding an intermediate INTERFACE library should not make a real
difference, and most online guides add a direct dependency between the
IMPORTED library and ExternalProject, we found this change to be a fix.

FindV8 is refactored a bit since otherwise adding the intermediate
library for downloaded V8 would otherwise be a bit ugly.