amun/simulator/simrobot: fix holdBall constraint
Commit d5d3db6a80 tried to fix the issue of the robots tipping over by trying
to set the joint position lower, but instead it set the relative z position of
the ball to the midpoint of the robot, so with simulate_dribble off the ball
floated in front of the robot, which among other things caused erroneous vision
coordinates, due to the projection.
This commit undoes this is and instead introduces a constraint that restricts
the rotation of a dribbling robot, preventing it from tipping over.
This might cause other issues, but then again so will every "solution" to this
problem, as it is inherently hacky.