glados/group/moves: make freekick moves ALLOW_EXTRA_ATTACKER=true
This includes having the moves pick the best fitting robot instead of using
the highest y coordinate, as was previously the case.
The main attacker for these moves are always sorted first, because every single
one of them made the first robot in their array the MA anyways and this way we
can now pick a more fitting MA depending on the move.