amun: processor: Factor out injecting and clearing of debug values
Before Tracker::worldState not only filled the world state, but also
filled in geometry and debug values of the Status message. For clarity,
it would be preferable to only fill in the world state. After this
commit, Tracker::worldState only fills in the first two, with geometry
being eliminated by a later commit.

In the meantime, this commit should make no noticable difference:
- For the speed and simple trackers, the debug values stored in the
  Status are simply discarded, since only parts of the generated Status
  are actually used. We still need to clear them to prevent a memory
  leak.
- For the normal tracker, the debug values were only added to the first
  Status packet (i.e. during the first call to assembleStatus), not the
  strategyStatus packet, since they were cleared during the first call
  to m_tracker->worldState().