core: timer: Always use a monotonic clock as system time
In contrast to QueryPerformanceCounter, both CLOCK_REALTIME and
gettimeofday are not monotonic. Calls performed later (in terms of wall
time) may return smaller values than earlier calls, for example due to
NTP or system administrator adjustments. This would lead to negative
time deltas, corrupting logfiles and causing issues in the tracking.
While this is unlikely to have happened often over the last 14 years, it
is preferrable to use an actual monotonic clock.

We use std::steady_clock instead of a custom implementation, since it is
more simple and offers a high precision clock on all modern systems (in
fact, it uses CLOCK_MONOTONIC on Posix systems in both libstdc++ and
libc++. The mingw-w64 clock_gettime CLOCK_MONOTONIC uses
QueryPerformanceCounter)

Note that while the underlying clock is now monotonic, the Timer itself
still is not, since it allows setting its time backwards. This is mostly
used for log replay and analysis, so it should not cause issues under
normal conditions.