amun: processor: Keep vision wrappers for Status in the Processor, not Tracker
In the future, this will enable passing only detections instead of
wrappers to the tracker. It also improves efficiency by a bit by not
copying the wrappers in the simple tracker only to discard them later
(as the vision part of its world state is not used).

This slightly changes behavior: prior to this commit, if the tracking
was reset via command, and there were any wrapper deemed older than the
reset time, all packets in m_visionPackets (i.e. those that still need
to be processed) were discarded. Since the m_visionPackets are added to
m_detectionWrappers at the beginning of the Tracker::process loop, and
m_detectionWrappers is eventually added to the world state, this leads
to some wrappers potentially missing from the status packet if the
tracking is reset.

After this commit, the status packet simply contains all vision packets
received during the current process tick. Since one of the purposes of
this feature is to archive vision packets, this is beneficial.