amun: processor/tracker: Introduce new vision latency calculation
This should allow an easier (and for the camera frame rate: automatic)
adaption of new latency values when something about the vision setup
changes.

Also rename a field in configdialog to ensure old configurations values
do not get loaded, as the user-settable delay does not model the whole
vision delay anymore.

Note that this calculation is wrong for the internal simulator, as it
does not simulate a shutter or transmission from camera to vision, and
has lower latency to Ra (since there is no network inbetween).