amun: processor: Statically calculate radio command delay
The value set in the configuration widget is now interpreted as a custom
value to be added on top.

The default value for the custom factor is 34, since before, we used a
radio delay of 45ms during games on the Robolaptop, and the new static
values add up to roughly 11ms.