ra: fieldwidget: fix virtual field aoi
The previous aoi assumed that the virtual transformation would transform the
topLeft and bottomRight of the real geometry to the topLeft and bottomRight of
the virtual field.
This is not the case. Instead it transforms it to the points where they are
relative to the new center and orientation of the virtual field.
This means that they could lie way outside or inside of the virtual field.