protobuf: Rename RobotPosition to TransformedRobotMeasurement
Its not actually only a position anymore and "RobotPosition" is fairly
generic. It is also not "raw" in the sense that these are in the vision
coordinate system. They are scaled from mm to m. Also, the transform is
applied (i.e. they are flipped and transformed for the virtual field if
enabled)