firmware: interface: add EPSILON in conversion to avoid ugly edge case
Previously, when mapping 0 from a symmetric interval [-x, x] to a N bit signed integer interval,
it resulted in -1, because it was computed by roundf(-.5f).

To just slightly tip the rounding towards +inf, we add EPSILON. That way, the call to
roundf has an argument slightly larger than -.5, which rounds to zero.