amun: tracking/robotfilter: Use global velocity given by the command evaluator
We sample these from the trajectory (since 090c0e57f833). Using them
directly (instead of using sampled local velocities) hopefully leads to
less rotation-state-error propagating to the velocity.

The filter normally handles data in "tracking" coordinates, i.e. those
sent out by the vision, but rotated by -90°. In tracking coordinates,
positive y points at one of the goals. Note that at this point it is
unspecified which goal.

The command evaluator instead outputs data in what we call "global"
coordinates, which are those of the yellow team (i.e. positive y points
at the blue goal, negative y points at the yellow goal)

The game controller has a flag "blueTeamOnPositiveHalf" (internally we
call this "flipped", although we set it to !blueTeamOnPositiveHalf).
When blueTeamOnPositiveHalf is true (isFlipped is false), the goal sides
for tracking and global coordinates are the same, as positive y points
at the blue goal in both cases. However, when the blue team is on the
negative half, they are inverse, leading to our tracking using wrongly
rotated coordinates.

We thus need to apply the field transform in RobotFilter::predict, as we
do in RobotFilter::get.