glados: t/s/duel: make duel more robust against rotate and shoot
This might make duel slightly worse for the past situations it was built for,
but hopefully also slightly better against e.g. TIGERs rotate and shoot.
The idea is, since the TIGERs bot can rotate with the ball so fast, that it
can easily lose the duelling robot if the duel tries to go for the ball.
So instead we replace the previous used future ball position to the midpoint
between the opponent robot and that future ball position to let the guarding
position be less affected by the rotation of the opponent robot.
Setting it directly to the midpoint of the robot causes the duel position to
be too far off from the ball to goal line, which makes it more dangerous.