cmake: Allow protobuf files in subdirectories
This is something we eventually want.

This requires the usage of protobuf_generate over protobuf_generate_cpp.
protobuf_generate is only available starting with CMake 3.13, however we
do not support any old Ubuntu versions where this would be a problem
(even 20.04 has support for 3.13)

We also need to move every protobuf file to a subdirectory called
protobuf. Since we want users of the library to be able to include the
generated .pb.h files using #include "protobuf/foo.pb.h", we need to
provide "protobuf" somewhere in their include path. This was previously
done by "reusing" the protobuf path containing the CMakeLists. However,
this would break for nested directory structures. The generated files
would not use the protobuf prefix (since protoc has never seen it),
causing included paths in generated files to be wrong. For non nested
structures, this was no problem, since the include would simply refer to
a file in the same directory.

Setting the protobuf_generate IMPORT_DIRS to
${CMAKE_CURRENT_BINARY_DIR}/.. instead (and thus reusing the existing
protobuf path) did not work.