protobuf: Mirror upstream folder structure for small-size-league protocols
That way, it is easier to see the correspondance between a local copy of
the files and their upstream version. This is relevant when checking for
diffs or updating them.

Some of the files are present in multiple source repositories. We group
them according to the technical overview page [1] into
- Vision
- Game Controller
- Simulation

Also we separate some old definitions, which are still present in old
logs but are deprecated for new usage, or relate to old SSL stuff.

[1] https://ssl.robocup.org/technical-overview-of-the-small-size-league/#protobuf-definitions