amun: Remove remnants of mixed team challenge
The sendMixedTeam function is simply removed from the Typescript
interface since I am pretty sure it has not been used once since
Typescript support was introduced.

In Lua, the function remains, although it errors, since it is a bit more
deeply woven into the Lua code and I do not want to invest the time
removing it.