amun: processor: Fix usage of has_vision_data
This fixes an oversight of dd57e6aaef39. Prior to that commit,
has_vision_data was set to true, if any vision wrapper was received
since the last tracking reset (which is how the strategy uses it). I
wrongfully assumed the meaning of the field is "received vision wrappers
during this tick"