amun: processor/worldparameters: Handle finishProcessing correctly
Due to an oversight in c717e1a32f38, m_ballModelUpdated was never reset
to false, leading to the geometry being included in every call to
getGeometryUpdate.

Also set geometryUpdated to true initially to ensure the geometry is
copied during the first frame. While this would also be the case due to
ballModelUpdated being initialized to true, this is more explicit.