amun: internalgamecontroller: Change division via command
Previously the InternalGameController would only tell the embedded CI GC
to change division when the division sent via world state changed.

The processor calculated the world state division from the geometry, and
we change the simulator setup and thus the geometry when changing the
division in the UI. However nowadays, the division on world state is
primarily calculated from the team info (max allowed robots etc.) which
is sent by the GC. Since the GC never received a command to change the
division, it still sent out the old team info.

Also rename MainWindow::changeDivision to MainWindow::changeFieldSetup
since it does not really change the division (and that was never what it
did primarily anyways)