ra: refereewidget: Allow toggling Game Controller auto continue
I assume part of the reason why the game controller is only optionally
enabled, is that it always used auto_continue. This can be a hinderance
when trying to test. Since a later commit will require the game
controller to run when using the internal referee, being able to toggle
auto continue is necessary.

Set it to false by default since this is what non-software people
probably expect when opening up Ra.