amun: Always run actual Game Controller when internal Referee is active
Otherwise, Ra has to keep implementing part of what the game controller
does. This can lead to bugs, where we do not update the team
info used to calculate the division, leaving us unable to change the
simulated division. By requiring the actual Game Controller, we use a
proven source of Referee messages.

I assume part of the reason why the game controller was only optionally
enabled, is that it always used auto_continue. This can be a hinderance
when testing. However, since HEAD~, it is possible to toggle this.

The flag also enabled and disabled the internal autoref. In order to do
that now (i.e. run with or without Autoref), the user needs to load or
disable to Autoref strategy in the TeamWidget.

We no longer need to forward Referee packets when the
InternalGameController is disabled, since we simply don't produce custom
Referee packets in this case.