amun/core/loganalyzer/simulator-cli/visionlog: Fix Protobuf ByteSize deprecation warning
ByteSize has been replaced by ByteSizeLong and should not be used
anymore. Since ByteSize was only used to resize QByteArrays, and
QByteArray::resize takes an int (qsizetype in Qt6), we check whether the
result of ByteSizeLong would be too large to be represented by that.

trackingreplay initialized QByteArrays to 0. This is not the case
anymore but should not matter, since the buffer is filled by Protobuf
immediately after.