amun/p/radiosystem: change error methods to all take or return TransceiverError
Previously the TransceiverErrors were taken apart and put together again in
multiple places and it makes no sense to not just propagate the whole error
as a single struct.
Also fixes a few places where the TransceiverError was not propagated correctly
despite it having the [[nodiscard]] attribute (which wasn't shown due to
compiler warnings being broken for a while).