ra: conditionally replace deprecated stateChanged with checkStateChanged
The checkStateChanged method is introduced in Qt6.7 where the signal type
changes from int to Qt::CheckState. This is a sensible change and in one
place we already casted the int to Qt::CheckState, but since for now we need
to support down to Qt6.2, this is only enabled with an #if version check.