amun: processor/tracker: Fix -Wmaybe-uninitialized
QMap::operator[] (same for std::map, std::unordered_map etc.) has to
default-construct a value if the map does not contain the value, which
in the case of Eigen, does not initialize its fields for performance
reasons.

The default constructed value seems to be read somewhere [1] so avoiding
the default constructed value via insert avoid the problem.

[1] Though I can't be bothered to completetly digest the error message
due to too many templates