tracking: ballgroundcollisionfilter: Pass in position of primary camera from the outside
Previously, the filter would lookup the position using the map stored in
the camera info every time it needed the position. Since this was done
using QMap::operator[], a default constructed vector was returned
whenever the map did not contain a position for the given camera id. The
default constructor for Eigen vectors leaves its fields uninitialized,
so this led to a lot of -Wmaybe-uninitialized triggering.

So we query the map on the outside and pass the position in through the
constructor instead, avoiding the warnings. As a bonus, this should be a
bit more performant. The check whether the position is available
shouldn't™ trigger since we check whether the camera has a position
before creating the filter, but better safe than sorry.

Technically, this behaves worse if the camera moves during the game, at
least until we next cross a camera boundary and recreate the filter. But
ex falsum quodlibet.