amun: internalgamecontroller: Set all events to accept
It was like this prior to ed3d5ecf5898 as this was set via engine.yml.
However we now regenerate the config on startup and have to set it
again.

Not accepting PLACEMENT_SUCCEEDED leads to the game stalling, even
though the robot successfully placed the ball and the autoref proposed
this.