amun: gamecontroller/gamecontrollerci: Make not finding a port an error condition
We don't detect this from the GC itself, as this is not an error for the
GC (it would only be one for the frontend, and we disable that), and we
don't want to parse stdout.

Also wrap the allocated ports in a struct, since we'll allocate more
later and pass them around.