amun: gamecontroller/gamecontrollermediator: Pass internal Autoref events using a standard socket
This more closely aligns the function of the internal autoref with that
used during actual games, which hopefully leads to the game controller
behaving more as expected. For example, we no longer need to set all
events to ACCEPT, since now, the game controller sees an actual single
Autoref being connected and having the majority by itself.