amun: trajectorypath: stabilize trajectory generation
The problem this fixes is that if the robot only needs to brake to reach
the goal (+/- a few centimeters), the success rate of findTrajectory
decreased dramatically.
This lead to the StandardSampler being used instead of the direct
trajectory, which often does not find the best trajectory, thus
resulting in overshooting.