amun: trajectorypath: fix acceleration in trajectory given to strategy
The problem is that the strategy needs to compute the acceleration for
the spline that is sent to the robot. This is currently done by sampling
the trajectory at 10ms in the future and computing the acceleration by
comparing it to the current time.
The problem occurred when the quadratic slowdown was used but the
trajectory segments sent to the strategy did not include them (since
they were too far in the future). Since the lower acceleration in the
slowdown segment was included in the last segment, once the robot
reached the point where it only needed to brake, the strategy
incorrectly computed a lower acceleration than intended.
This lead to the robot not being able to brake as intended, thus
overshooting.

The problem is fixed by adding an additional segment at the slowdown
start time.