amun: typescript: Use unique_ptr for require cache
v8::Global (also known as v8::UniquePersistent) manages the lifetime of
objects on the V8 heap using C++ constructors and destructors. Thus, if
we forget to free the allocated pointer, we leak both the memory the
pointer points to and the memory on the V8 heap.

Since this can easily happen, we instead use unique_ptr. It would also
be possible to use the value directly, without extra heap allocation.
However, v8::Global's move constructor leads to errors in valgrind - it
seems to depend on uninitialized values. This can probably be fixed
through a V8 update but in the meantime we use unique_ptr instead.

This requires the usage of std types instead of their Qt counterparts,
as the latter don't handle move semantics and v8::Global is move-only.