amun: typescript: Add RAII wrapper for v8::Isolate
This simplifies the depending code a bit, and its good C++ practice. It
also allows us to remove a few manual deinitializations from ~Typescript
since they now happen implicitly before the Isolate is disposed.

We keep a non-owning pointer in m_isolate in order to not have to adjust
every second line.

Initialization/Deinitialization order may be different (since the
Isolate is constructed implicitly) but this should not matter. Either
the other constructed objects do not depend on the isolate and their
init order does not matter, or they depend on the isolate, in which case
the isolate should be initialized before and destroyed after the other
object.

We don't use it in tsc_internal for now since it has more complicated
usage of Isolate scoping.