amun: typescript: Delete QTPath instances on reload
Previously, the QTPath instances used the QT ownership mechanism (via
parent pointer) to be deleted when Typescript deconstructs. The
Typescript strategy is reloaded in place, so this only happens when we
close Ra or disable the Strategy. Thus, the instances start to pile up.

We now clear them explicitly when loading the strategy instead. Since we
store the instances in unique_ptr, QTPath no longer has to be a QObject.
The instances are still deleted when Typescript deconstructs as the
containing vector deconstructs as well.