amun: js_path: Delete destroy method
The method was a footgun that allowed the strategy to cause
double-frees. The returned and deleted pointer was copied out of a
unique_ptr, which is also deleted when the Typescript instance is
deleted (since QTPath is a child of Typescript)

We simply delete it since it has never been used during Glados's
existence. Replayability of old strategy versions is thus not hindered.