base: coordinates: normalized angle to -pi to pi
Previously, the Invert coordinates type assumed that the input
angle was normalized to [0, 2pi], and it respected that - if
the input was in that range, the output was as well.

Sadly, the normalization to [-pi, pi] is much more widely used
in the strategy, and also in Ra/Amun. For example, this unconventional
choice of normalization lead to the following effect:
When setting the robot direction, the value received from Amun is
normalized to [-pi, pi], but in global coordinates.
This value was then passed to Coordinates.toLocal, and if the
team was blue, this lead to the robot direction being in the range of
[0, 2pi].

For the yellow team however, the Pass coordinates type just returned the
input value (normalized to [-pi, pi] from Amun).

So for team blue, robot directions were in [0, 2pi], and in [-pi, pi]
for yellow. This commit normalizes to [-pi, pi] no matter them team
color.