cmake: Remove double variable expansion from if
CMake if already expands variables by itself, so if the value contained
in one of the variables were itself a variable name, it would be
expanded twice. Since this can lead to hard to spot bugs, we remove the
first expansion.

As an added bonus, this works a bit better for comparisions if the
variable happens to be undefined.