amun: tracking: Fix compiler warning
The compiler throught, cam in calcDistToCamera may have been
uninitialized since the value default constructs its values if the key
is not present in the map, and the Eigen default constructors don't
initialize fields.

The check whether the position is available shouldn't™ trigger since we
check whether the camera has a position before creating the filter, but
better safe than sorry.

Similarly to 35c46cc8b67e, we now pass in the position from the outside.