cmake: protobuf: Update to v4.25.7
Download via Git for now since there will be another update in the near
future and we don't want to maintain a tarball on our server for this
intermediate version.

This requires a lot of additional CMake code since Protobuf C++ 4
introduced a dependency on abseil.

The update also introduced a warning about an unused return value, which
is now fixed.