cmake: protobuf: Update to v5.29.4
Arena::CreateArena was deprecated in favor of Arena::Create and will be
removed in v30 [1].

always_print_primitive_fields was removed as well. The docs are not
exactly clear on the replacement [2], but it seems to be an effort to
handle output of default-valued fields consistently accross proto2 and
proto3.

[1] https://protobuf.dev/news/v30/ "Remove Deprecated APIs"
[2] https://protobuf.dev/news/v26/ "JSON Formatter Option Changes"