cmake: protobuf: Always build from source
Searching for version 3.21.12 is no longer correct since the last
Protobuf update, we now use version 5.29.4. The top level CMake code was
not adjusted for that.

We could in principle continue using the system version, but we'd need
to...
- handle the case were the installed version is too new and thus
  breaking
- configure the proper dependencies (absl, etc.) for the system version,
  since the module mode find_package script does not do that, or use the
  config mode script
- integrate that with cross compilation and the lib::protobuf target

Which seems relatively complicated, and we'd still have to build from
source on most systems anyway, since the installed version isn't a match
most of the time. So we simply always build from source.

Also sneakily update the version variable.