ra: improve virtual field ui
Add preview of the virtual geometry in the setup dialog.
Update the short cuts for the dimensions to use division A and B.
The "quad size"/"double size" convention has long fallen out of favor and most
current members don't know what they mean.
It's also more practical, since those are the standard field sizes and Div B is
exactly half of a Div A field.