amun: processor/radiosystem: dont write to other transceivers after one reported an error
When two HBC transceivers are connected, and the first one in the
transceivers list reported an error, the transceiver list was
immediately cleared, and then the second transceivers was used.

This obviously resulted in a crash an is easily fixed by simply no
using the second transceiver anymore after the first one raised an
error.