ra: fieldwidget: fix dragging robots when using the field
I dont fully understand, why the flipping is only necessary when the
virtual field is active, it doesnt make sense, but I thoroughly tested
all possible combinations of team color, virtual field and flipped. So
I confindent to say that this works.

In the future one might want to think through all the transformations
and make it make sense.